# backend/docs/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrainerPlatform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/domains/{domain_slug}": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Get Domain",
        "description": "Return a domain detail including its published trainers.",
        "operationId": "get_domain_api_v1_domains__domain_slug__get",
        "parameters": [
          {
            "name": "domain_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/certification-core/items": {
      "get": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "List Items",
        "operationId": "list_items_api_v1_certification_core_items_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          },
          {
            "name": "domain_pack_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Domain Pack Id"
            }
          },
          {
            "name": "item_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Item Type"
            }
          },
          {
            "name": "item_family_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Item Family Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "Create Item",
        "operationId": "create_item_api_v1_certification_core_items_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/certification-core/items/{item_id}": {
      "get": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "Get Item",
        "operationId": "get_item_api_v1_certification_core_items__item_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "Update Item",
        "operationId": "update_item_api_v1_certification_core_items__item_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/certification-core/domain-packs": {
      "get": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "List Domain Packs",
        "operationId": "list_domain_packs_api_v1_certification_core_domain_packs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          },
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Locale"
            }
          },
          {
            "name": "market",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Market"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainPackListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "Create Domain Pack",
        "operationId": "create_domain_pack_api_v1_certification_core_domain_packs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainPackCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainPackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/certification-core/domain-packs/{domain_pack_id}": {
      "get": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "Get Domain Pack",
        "operationId": "get_domain_pack_api_v1_certification_core_domain_packs__domain_pack_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "domain_pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainPackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "Update Domain Pack",
        "operationId": "update_domain_pack_api_v1_certification_core_domain_packs__domain_pack_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "domain_pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain Pack Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainPackUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainPackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DomainDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "trainers": {
            "items": {
              "$ref": "#/components/schemas/TrainerSummaryResponse"
            },
            "type": "array",
            "title": "Trainers"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "icon",
          "trainers"
        ],
        "title": "DomainDetailResponse",
        "description": "Domain detail including its published trainers."
      },
      "DomainPackCreate": {
        "properties": {
          "domain_pack_id": {
            "type": "string",
            "maxLength": 100,
            "title": "Domain Pack Id"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "version": {
            "type": "string",
            "maxLength": 20,
            "title": "Version"
          },
          "locale": {
            "type": "string",
            "title": "Locale",
            "default": "en-US"
          },
          "market": {
            "type": "string",
            "title": "Market",
            "default": "global"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "competency_framework_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competency Framework Id"
          },
          "blueprint_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blueprint Ids"
          },
          "knowledge_source_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Source Ids"
          },
          "item_bank_policy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Bank Policy Id"
          },
          "scoring_policy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scoring Policy Id"
          },
          "pass_policy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pass Policy Id"
          },
          "rubric_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rubric Ids"
          },
          "supported_modes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supported Modes"
          },
          "created_by": {
            "type": "string",
            "maxLength": 100,
            "title": "Created By"
          }
        },
        "type": "object",
        "required": [
          "domain_pack_id",
          "name",
          "version",
          "created_by"
        ],
        "title": "DomainPackCreate"
      },
      "DomainPackListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DomainPackResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 100
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "DomainPackListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemCreate": {
        "properties": {
          "item_id": {
            "type": "string",
            "maxLength": 100,
            "title": "Item Id"
          },
          "item_family_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Family Id"
          },
          "domain_pack_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Pack Id"
          },
          "item_type": {
            "type": "string",
            "maxLength": 50,
            "title": "Item Type"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "response_contract": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Contract"
          },
          "answer_key": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Key"
          },
          "rubric_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rubric Id"
          },
          "competency_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competency Ids"
          },
          "knowledge_source_refs": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Source Refs"
          },
          "difficulty_target": {
            "type": "string",
            "title": "Difficulty Target",
            "default": "medium"
          },
          "locale": {
            "type": "string",
            "title": "Locale",
            "default": "en-US"
          },
          "market": {
            "type": "string",
            "title": "Market",
            "default": "global"
          },
          "created_by": {
            "type": "string",
            "maxLength": 100,
            "title": "Created By"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "item_type",
          "created_by"
        ],
        "title": "ItemCreate"
      },
...
      "ItemUpdate": {
        "properties": {
          "prompt": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "response_contract": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Contract"
          },
          "answer_key": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Key"
          },
          "rubric_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rubric Id"
          },
          "competency_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competency Ids"
          },
          "knowledge_source_refs": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Source Refs"
          },
          "difficulty_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty Target"
          },
          "difficulty_measured": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty Measured"
          },
          "discrimination_measured": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discrimination Measured"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "title": "ItemUpdate"
      },
...
      "TrainerSummaryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "trainer_product_id": {
            "type": "string",
            "title": "Trainer Product Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "product_type": {
            "type": "string",
            "title": "Product Type"
          }
        },
        "type": "object",
        "required": [
          "id",
          "trainer_product_id",
          "slug",
          "name",
          "description",
          "product_type"
        ],
        "title": "TrainerSummaryResponse",
        "description": "Minimal trainer representation used in domain detail responses."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}